refactor: #222 reorganize project layout
